+++ CHANGELOG.md
@@ -34,6 +34,7 @@
 - [#437] - Keep SSG silent when using `0Hxy` in a muted track (thanks [@Zexxerd])
 - [#443] - Fix song length estimation to handle tempo/speed change correctly (thanks [@Drillimation])
 - [#444] - Fix unexpected envelope reset of FM3-OP4 (thanks [@Zexxerd])
+- [#448] - Check if the read/write position of the binary container is valid (thanks [@OPNA2608])
 
 [@cxong]: https://github.com/cxong
 
@@ -67,6 +68,7 @@
 [#437]: https://github.com/BambooTracker/BambooTracker/issues/437
 [#443]: https://github.com/BambooTracker/BambooTracker/issues/443
 [#444]: https://github.com/BambooTracker/BambooTracker/issues/444
+[#447]: https://github.com/BambooTracker/BambooTracker/issues/447
 
 ## v0.5.0 (2021-08-01)
 ### Added